aclocal.m4: Update AC_DEFINE usage.
Since autoconf 2.69 all AC_DEFINE parameters are required.

"If neither value nor description are given, then value defaults to 1
instead of to the empty string. This is for backwards compatibility with
older versions of Autoconf, but this usage is obsolescent and may be
withdrawn in future versions of Autoconf."
(https://www.gnu.org/savannah-checkouts/gnu/autoconf/manual/autoconf-2.69/html_node/Defining-Symbols.html)